No longer double/triple buffering each filter. I think this is fine to do

+++ src/rendering/backend/renderer_fbo.cpp
@@ -86,7 +86,7 @@ Result FramebufferObject::Create(Device *device, RenderPass *render_pass)
     VkFramebufferCreateInfo framebuffer_create_info{};
     framebuffer_create_info.sType = VK_STRUCTURE_TYPE_FRAMEBUFFER_CREATE_INFO;
     framebuffer_create_info.renderPass = render_pass->GetRenderPass();
-    framebuffer_create_info.attachmentCount = attachment_image_views.size();
+    framebuffer_create_info.attachmentCount = uint32_t(attachment_image_views.size());
     framebuffer_create_info.pAttachments    = attachment_image_views.data();
     framebuffer_create_info.width = m_width;
     framebuffer_create_info.height = m_height;
